Enemy, Drill, Sliding
Worked on the ability to collect energy when drilling, and fixed the Enemy behavior script, and fixed the sliding event that happens when the player stays still.

+++ Assets/Scripts/DrillEnergy.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class DrillEnergy : MonoBehaviour {
 
     public int startingenergy = 0;
-    public int OurEnergy;
+    public float OurEnergy;
     public Slider EnergySlider;
     public Image FillImage;
     public Color m_FullHealthColor = Color.green;
